code/SiLabs/sensor_data_receiver/requirements.txt
bluepy==1.3.0
paho-mqtt==1.6.1
